refactor: move components aroud
